backend/python/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
websockets>=11.0.2
httpx>=0.24.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.3
supabase>=1.0.3
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
ccxt>=4.0.0
